Little reorganization
