+++ Assets/Prefabs/Managing/_GameManager.prefab
@@ -61,7 +61,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 8712221511988029991}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 38cf37b4f03c2cb4093e5df52ef4dcc8, type: 3}
+  m_Script: {fileID: 11500000, guid: 821be0c8ba792ee438dd33762463d421, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   dontDestroyOnLoad: 1
@@ -81,6 +81,7 @@ MonoBehaviour:
   playerSpawnMethod: 0
   spawnPrefabs: []
   gameManager: {fileID: 4353379490438976488}
+  HUD: {fileID: 1483694915}
 --- !u!114 &2722749715747069115
 MonoBehaviour:
   m_ObjectHideFlags: 0
