+++ Assets/Scripts/Management/Game/GameSettings.cs.meta
