+++ Assets/Scripts/Management/Network/ServerManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4147e27ae4d2a0a4baa2756ce0452392
+guid: 821be0c8ba792ee438dd33762463d421
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
